+++ packages/block/test/from-rpc.spec.ts
@@ -3,7 +3,14 @@ import * as tape from 'tape'
 
 import { blockFromRpc } from '../src/from-rpc'
 import { blockHeaderFromRpc } from '../src/header-from-rpc'
+import { Block } from '../src/index'
 
+import { MockProvider } from './mockProvider'
+import * as alchemy14151203 from './testdata/alchemy14151203.json'
+import * as infura15571241woTxs from './testdata/infura15571241.json'
+import * as infura15571241wTxs from './testdata/infura15571241wtxns.json'
+import * as infura2000004woTxs from './testdata/infura2000004wotxns.json'
+import * as infura2000004wTxs from './testdata/infura2000004wtxs.json'
 import * as blockDataDifficultyAsInteger from './testdata/testdata-from-rpc-difficulty-as-integer.json'
 import * as testDataFromRpcGoerliLondon from './testdata/testdata-from-rpc-goerli-london.json'
 import * as blockDataWithUncles from './testdata/testdata-from-rpc-with-uncles.json'
@@ -104,4 +111,67 @@ tape('[fromRPC]:', function (t) {
     st.equal(`0x${block.hash().toString('hex')}`, testDataFromRpcGoerliLondon.hash)
     st.end()
   })
+})
+
+tape('[fromRPC] - Alchemy/Infura API block responses', (t) => {
+  t.test('should create pre merge block from Alchemy API response to eth_getBlockByHash', (st) => {
+    const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.London })
+    const block = blockFromRpc(alchemy14151203, [], { common })
+    st.equal(`0x${block.hash().toString('hex')}`, alchemy14151203.hash)
+    st.end()
+  })
+
+  t.test(
+    'should create pre and post merge blocks from Infura API responses to eth_getBlockByHash and eth_getBlockByNumber',
+    (st) => {
+      const common = new Common({ chain: Chain.Mainnet })
+      let block = blockFromRpc(infura2000004woTxs, [], { common, hardforkByBlockNumber: true })
+      st.equal(
+        `0x${block.hash().toString('hex')}`,
+        infura2000004woTxs.hash,
+        'created premerge block w/o txns'
+      )
+      block = blockFromRpc(infura2000004wTxs, [], { common, hardforkByBlockNumber: true })
+      st.equal(
+        `0x${block.hash().toString('hex')}`,
+        infura2000004wTxs.hash,
+        'created premerge block with txns'
+      )
+      block = blockFromRpc(infura15571241woTxs, [], {
+        common,
+        hardforkByTTD: 58750000000000000000000n,
+      })
+      st.equal(
+        `0x${block.hash().toString('hex')}`,
+        infura15571241woTxs.hash,
+        'created post merge block without txns'
+      )
+
+      block = blockFromRpc(infura15571241wTxs, [], {
+        common,
+        hardforkByTTD: 58750000000000000000000n,
+      })
+      st.equal(
+        `0x${block.hash().toString('hex')}`,
+        infura15571241wTxs.hash,
+        'created post merge block with txns'
+      )
+
+      st.end()
+    }
+  )
+  t.end()
+})
+
+tape('[fromEthersProvider]', async (t) => {
+  const common = new Common({ chain: Chain.Mainnet, hardfork: Hardfork.London })
+  const provider = new MockProvider()
+  const blockHash = '0x1850b014065b23d804ecf71a8a4691d076ca87c2e6fb8fe81ee20a4d8e884c24'
+  const block = await Block.fromEthersProvider(provider, blockHash, { common })
+  t.equal(
+    '0x' + block.hash().toString('hex'),
+    blockHash,
+    'assembled a block from blockdata from a provider'
+  )
+  t.end()
 })